pyproject.toml
[project]
name = "arkiv-sdk"
version = "0.1.0"
description = "Python SDK for Golem DB networks - Web3.py + Entities"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "web3==7.13.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
# Testing infrastructure
test = [
    "pytest>=8.4.2",
    "testcontainers==4.13.1",
    "websockets>=13.0",
]

# Code quality and linting
lint = [
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]

# Development utilities
dev = [
    "python-dotenv>=1.0.0",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore:The @wait_container_is_ready decorator is deprecated.*:DeprecationWarning:testcontainers.*",
    "ignore:websockets.legacy is deprecated;.*",
]

[tool.mypy]
# Strict type checking configuration
strict = true
warn_return_any = true
warn_unused_configs = true
no_implicit_reexport = true
show_error_codes = true
show_column_numbers = true

# Per-module options
[[tool.mypy.overrides]]
module = [
    "testcontainers.*",
    "websockets.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
